ASM-7793 Create symlink to most recent CompCU.jar
The dell-compellent module includes several versions of the CompCU jar
file. Previously code changes had to be made whenever a new version of
the jar was added. This change automatically creates a symlink in the
latest dell-compellent rpm pointing to the newest version of the
CompCU jar file.

Separate code changes will be made to the dell-compellent module and
asm-core java code to always use the symlinked jar file to ensure they
always get the newest version of the jar.